+++ install_module_template/generic/action.sh
@@ -47,10 +47,7 @@ else
         key_check
         if [[ "$keycheck" == "KEY_VOLUMEUP" ]]; then
             echo "- 正在为你安装KSU Web UI，请稍等~"
-            unzip -jo "$ZIPFILE" 'common/apks/KsuWebUI.apk' -d /data/local/tmp/ &>/dev/null
-            pm install -r /data/local/tmp/KsuWebUI.apk &>/dev/null
-            rm -rf /data/local/tmp/KsuWebUI.apk
-            rm -rf "$MODPATH"/common/apks/KsuWebUI.apk
+            pm install -r "$MODPATH"/common/apks/KsuWebUI.apk &>/dev/null
             HAS_BEEN_INSTALLED_KsuWebUI_APK=$(pm list packages | grep io.github.a13e300.ksuwebui)
             if [[ $HAS_BEEN_INSTALLED_KsuWebUI_APK == *"package:io.github.a13e300.ksuwebui"* ]]; then
                 echo "- 好诶，KSUWebUI安装完成！"